Force dark mode when switch is activated
