+++ theme/assets/src/front-end/index.js
@@ -25,7 +25,7 @@ import { tabBarInit } from './components/tab-bar';
 import { embedsInit } from './components/embeds';
 import { widgetsInit } from './components/widgets';
 import { masonryInit } from './components/masonry';
-import { initDarkMode } from './dark-mode';
+import { initDarkModeSwitch } from './components/dark-mode-switch';
 import './components/navigation';
 import './components/skip-link-focus-fix';
 
@@ -46,5 +46,5 @@ document.addEventListener( 'DOMContentLoaded', () => {
 	embedsInit();
 	widgetsInit();
 	masonryInit();
-	initDarkMode();
+	initDarkModeSwitch();
 } );
